requirements.txt
binance-futures-connector
python-dotenv
colorlog